Authorization using Gate

+++ app/Http/Controllers/QuestionController.php
@@ -63,7 +63,16 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(401);
+        }
         return view('questions.edit', compact('question'));
+
+        // Cara 2
+        // if (\Gate::allows('update-question', $question)) {
+        //     return view('questions.edit', compact('question'));
+        // }
+        // abort(401);
     }
 
     /**
@@ -75,6 +84,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(401);
+        }
+
         $question->update($request->validated());
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
     }
@@ -87,6 +100,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(401);
+        }
+
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
     }
